feat: add hourly wind/hum/cloud, start aggregator
